ranking/src/requirements.txt
kafka-python
urllib3